Multiple changes, read description.
Campos variable to simulate a camera, the objects move around this variable. When the campos variable is above 0 it will move all the blocks in the world left and right if it is negative.

Fixed a clipping issue where the player would go through the floor.

Added jumping animation.

Fixed holding space to infinitely jump.

Added some new open source textures.