Major Cleanup, Read Description
Created separate files for each class.

Created levels folder with JSON files, these files will be used to create the levels. In these files you specify the level name, id, player textures, level data and textures. Bit 0 will be the first texture and bit 1 will be the second etc.

When you run the level, if you run far enough in the first level it will switch to the next level to show how the level switching is done.